refactor(core): moves type guards to formats/codewalker
