+++ src/core/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './joaat';
 export * from './int32';
 export * from './bit';
-export * from './typeGuards';
